moved loop one down
