+++ services/web/server/tests/conftest.py
@@ -3,6 +3,7 @@
 # pylint: disable=unused-argument
 # pylint: disable=unused-variable
 
+import asyncio
 import json
 import logging
 import random
@@ -478,3 +479,11 @@ def mock_dynamic_scheduler(mocker: MockerFixture) -> None:
         "simcore_service_webserver.dynamic_scheduler.api.update_projects_networks",
         autospec=True,
     )
+
+
+@pytest.fixture
+async def loop(
+    event_loop: asyncio.AbstractEventLoop,
+) -> asyncio.AbstractEventLoop:
+    """Override the event loop inside pytest-aiohttp with the one from pytest-asyncio."""
+    return event_loop
